make RxJava helpers optional in Selector
