+++ library/src/main/java/com/github/gfx/android/orma/Selector.java
@@ -33,12 +33,6 @@ import android.support.annotation.RestrictTo;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import java.util.concurrent.Callable;
-
-import io.reactivex.Observable;
-import io.reactivex.ObservableEmitter;
-import io.reactivex.ObservableOnSubscribe;
-import io.reactivex.Single;
 
 @SuppressLint("Assert")
 public abstract class Selector<Model, S extends Selector<Model, ?>>
@@ -69,7 +63,6 @@ public abstract class Selector<Model, S extends Selector<Model, ?>>
         super(relation);
         orderBy = relation.buildOrderingTerms();
     }
-
 
     public Selector(@NonNull Selector<Model, ?> selector) {
         super(selector);
@@ -208,17 +201,6 @@ public abstract class Selector<Model, S extends Selector<Model, ?>>
         return (int) conn.rawQueryForLong(sql, getBindArgs());
     }
 
-    @CheckResult
-    @NonNull
-    public Single<Integer> countAsSingle() {
-        return Single.fromCallable(new Callable<Integer>() {
-            @Override
-            public Integer call() throws Exception {
-                return count();
-            }
-        });
-    }
-
     public boolean isEmpty() {
         return count() == 0;
     }
@@ -266,24 +248,6 @@ public abstract class Selector<Model, S extends Selector<Model, ?>>
             cursor.close();
         }
         return result;
-    }
-
-    @NonNull
-    public <T> Observable<T> pluckAsObservable(final ColumnDef<Model, T> column) {
-        return Observable.create(new ObservableOnSubscribe<T>() {
-            @Override
-            public void subscribe(ObservableEmitter<T> emitter) throws Exception {
-                Cursor cursor = executeWithColumns(column.getQualifiedName());
-                try {
-                    for (int pos = 0; !emitter.isDisposed() && cursor.moveToPosition(pos); pos++) {
-                        emitter.onNext(column.getFromCursor(conn, cursor, 0));
-                    }
-                } finally {
-                    cursor.close();
-                }
-                emitter.onComplete();
-            }
-        });
     }
 
     @CheckResult
@@ -340,24 +304,6 @@ public abstract class Selector<Model, S extends Selector<Model, ?>>
     @NonNull
     public Model newModelFromCursor(@NonNull Cursor cursor) {
         return getSchema().newModelFromCursor(conn, cursor, 0);
-    }
-
-    @NonNull
-    public Observable<Model> executeAsObservable() {
-        return Observable.create(new ObservableOnSubscribe<Model>() {
-            @Override
-            public void subscribe(ObservableEmitter<Model> emitter) throws Exception {
-                final Cursor cursor = execute();
-                try {
-                    for (int pos = 0; !emitter.isDisposed() && cursor.moveToPosition(pos); pos++) {
-                        emitter.onNext(newModelFromCursor(cursor));
-                    }
-                } finally {
-                    cursor.close();
-                }
-                emitter.onComplete();
-            }
-        });
     }
 
     // implements Iterable<Model>
